+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.27.0
 
   * Add PHP 8.3 polyfill for `json_validate()`
+  * Add polyfill for PHP 8.2's `Random` interfaces, exceptions and `Secure` engine
   * Fix `IntlDateFormatter::formatObject()` signature
 
 # 1.26.0